fix(drive)!: version the where-clause grammar and harden the multi-IN lowering per review
Addresses the four blocking review findings on the multi-IN PR:

- The where-clause grouping is now versioned like the lowering
  (`where_clause_grouping` feature version, flipping with
  `non_primary_key_path_query`): the query constructors take the
  platform version and dispatch to a v0 grouping restored verbatim from
  the pre-multi-IN implementation — so protocol version 13 keeps
  reporting MultipleInClauses for every multi-IN shape, ahead of
  duplicate-field, overlap, and range-grouping checks, exactly as
  historical nodes did — or to the v1 grouping that groups multiple IN
  clauses structurally.
- A versioned shape preflight runs at the top of path-query
  construction (both the server and verify paths, and the
  start-at-document proof verifier), so v0 rejects multi-IN and v1
  rejects multi-IN + cursor before the startAfter document is fetched
  from storage or any proof work happens; a nonexistent cursor can no
  longer surface as StartDocumentNotFound ahead of the shape error.
- The v1 multi-IN lowering now honors orderBy direction on left-over
  index properties through its own recursion (the shared v0 helper
  builds those levels with the index property's direction), with unit
  and execution + proof coverage for a descending left-over level.
- The v0 and v1 lowerings and their helpers moved out of query/mod.rs
  into query/non_primary_key_path_query/{v0,v1}, and the grouping
  implementations live in query/where_clause_grouping/{v0,v1}, keeping
  only dispatch at the parent boundary so live behavior is isolated
  from later edits.

Also from the review: the withdrawal transaction-index query now
carries its IN clause in in_clauses (identical lowering; it previously
worked only because to_path_query dispatches on the operator), and the
wasm-drive-verify parsers reject non-array in_clauses values instead of
letting Array::from coerce them to an empty array.
