+++ packages/rs-drive-abci/src/execution/platform_events/protocol_upgrade/perform_events_on_first_block_of_protocol_change/v0/mod.rs
@@ -1327,6 +1327,7 @@ mod tests {
             "select * from profile",
             &dashpay_v2_fetch_info.contract,
             Some(&platform.config.drive),
+            PlatformVersion::get(14).expect("expected platform version 14"),
         )
         .expect("expected a document query");
         let query_results = platform
