+++ packages/rs-drive/src/drive/document/query/query_documents_with_flags/v0/mod.rs
@@ -138,9 +138,13 @@ mod tests {
         let platform_version = PlatformVersion::latest();
 
         let sql = "select * from contactRequest";
-        let query =
-            DriveDocumentQuery::from_sql_expr(sql, &contract, Some(&DriveConfig::default()))
-                .expect("valid query");
+        let query = DriveDocumentQuery::from_sql_expr(
+            sql,
+            &contract,
+            Some(&DriveConfig::default()),
+            platform_version,
+        )
+        .expect("valid query");
 
         let outcome = drive
             .query_documents_with_flags_v0(query, None, true, None, platform_version)
@@ -159,9 +163,13 @@ mod tests {
         let platform_version = PlatformVersion::latest();
 
         let sql = "select * from contactRequest";
-        let query =
-            DriveDocumentQuery::from_sql_expr(sql, &contract, Some(&DriveConfig::default()))
-                .expect("valid query");
+        let query = DriveDocumentQuery::from_sql_expr(
+            sql,
+            &contract,
+            Some(&DriveConfig::default()),
+            platform_version,
+        )
+        .expect("valid query");
 
         let outcome = drive
             .query_documents_with_flags_v0(query, None, false, None, platform_version)
