+++ packages/rs-drive/src/query/mod.rs
@@ -82,10 +82,10 @@ use {
         data_contract::{
             accessors::v0::DataContractV0Getters,
             document_type::{accessors::DocumentTypeV0Getters, methods::DocumentTypeV0Methods},
-            document_type::{DocumentTypeRef, Index, IndexProperty},
+            document_type::{DocumentTypeRef, Index},
             DataContract,
         },
-        document::{document_methods::DocumentMethodsV0, Document, DocumentV0Getters},
+        document::{document_methods::DocumentMethodsV0, Document},
         platform_value::{btreemap_extensions::BTreeValueRemoveFromMapHelper, Value},
         version::PlatformVersion,
         ProtocolError,
@@ -132,12 +132,15 @@ pub mod conditions;
 mod defaults;
 #[cfg(any(feature = "server", feature = "verify"))]
 pub mod having;
+mod non_primary_key_path_query;
 #[cfg(any(feature = "server", feature = "verify"))]
 pub mod ordering;
 #[cfg(any(feature = "server", feature = "verify"))]
 pub mod projection;
 #[cfg(any(feature = "server", feature = "verify"))]
 mod single_document_drive_query;
+/// Versioned grouping of raw where clauses into equality / range / in buckets
+pub(crate) mod where_clause_grouping;
 
 // Module declarations exclusively for "server" feature
 #[cfg(feature = "server")]
@@ -341,7 +344,10 @@ impl InternalClauses {
 
     #[cfg(any(feature = "server", feature = "verify"))]
     /// Extracts the `WhereClause`s and returns them as type `InternalClauses`.
-    pub fn extract_from_clauses(all_where_clauses: Vec<WhereClause>) -> Result<Self, Error> {
+    pub fn extract_from_clauses(
+        all_where_clauses: Vec<WhereClause>,
+        platform_version: &PlatformVersion,
+    ) -> Result<Self, Error> {
         let primary_key_equal_clauses_array = all_where_clauses
             .iter()
             .filter_map(|where_clause| match where_clause.operator {
@@ -365,7 +371,7 @@ impl InternalClauses {
             .collect::<Vec<WhereClause>>();
 
         let (equal_clauses, range_clause, in_clauses) =
-            WhereClause::group_clauses(&all_where_clauses)?;
+            WhereClause::group_clauses(&all_where_clauses, platform_version)?;
 
         let primary_key_equal_clause = match primary_key_equal_clauses_array.len() {
             0 => Ok(None),
@@ -656,13 +662,20 @@ impl<'a> DriveDocumentQuery<'a> {
         contract: &'a DataContract,
         document_type: DocumentTypeRef<'a>,
         config: &DriveConfig,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         let query_document_value: Value = ciborium::de::from_reader(query_cbor).map_err(|_| {
             Error::Query(QuerySyntaxError::DeserializationError(
                 "unable to decode query from cbor".to_string(),
             ))
         })?;
-        Self::from_value(query_document_value, contract, document_type, config)
+        Self::from_value(
+            query_document_value,
+            contract,
+            document_type,
+            config,
+            platform_version,
+        )
     }
 
     #[cfg(any(feature = "server", feature = "verify"))]
@@ -672,9 +685,16 @@ impl<'a> DriveDocumentQuery<'a> {
         contract: &'a DataContract,
         document_type: DocumentTypeRef<'a>,
         config: &DriveConfig,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         let query_document: BTreeMap<String, Value> = query_value.into_btree_string_map()?;
-        Self::from_btree_map_value(query_document, contract, document_type, config)
+        Self::from_btree_map_value(
+            query_document,
+            contract,
+            document_type,
+            config,
+            platform_version,
+        )
     }
 
     #[cfg(any(feature = "server", feature = "verify"))]
@@ -684,6 +704,7 @@ impl<'a> DriveDocumentQuery<'a> {
         contract: &'a DataContract,
         document_type: DocumentTypeRef<'a>,
         config: &DriveConfig,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         if let Some(contract_id) = query_document
             .remove_optional_identifier("contract_id")
@@ -762,7 +783,8 @@ impl<'a> DriveDocumentQuery<'a> {
                     }
                 })?;
 
-        let internal_clauses = InternalClauses::extract_from_clauses(all_where_clauses)?;
+        let internal_clauses =
+            InternalClauses::extract_from_clauses(all_where_clauses, platform_version)?;
 
         let start_at_option = query_document.remove("startAt");
         let start_after_option = query_document.remove("startAfter");
@@ -855,6 +877,7 @@ impl<'a> DriveDocumentQuery<'a> {
         contract: &'a DataContract,
         document_type: DocumentTypeRef<'a>,
         config: &DriveConfig,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         let all_where_clauses: Vec<WhereClause> = match where_clause {
             Value::Null => Ok(vec![]),
@@ -916,6 +939,7 @@ impl<'a> DriveDocumentQuery<'a> {
             contract,
             document_type,
             config,
+            platform_version,
         )
     }
 
@@ -950,6 +974,7 @@ impl<'a> DriveDocumentQuery<'a> {
         contract: &'a DataContract,
         document_type: DocumentTypeRef<'a>,
         config: &DriveConfig,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         let limit = maybe_limit
             .map_or(Some(config.default_query_limit), |limit_value| {
@@ -964,7 +989,8 @@ impl<'a> DriveDocumentQuery<'a> {
                 config.max_query_limit
             ))))?;
 
-        let internal_clauses = InternalClauses::extract_from_clauses(where_clauses)?;
+        let internal_clauses =
+            InternalClauses::extract_from_clauses(where_clauses, platform_version)?;
 
         let order_by: IndexMap<String, OrderClause> = order_by_clauses
             .into_iter()
@@ -990,6 +1016,7 @@ impl<'a> DriveDocumentQuery<'a> {
         sql_string: &str,
         contract: &'a DataContract,
         config: Option<&DriveConfig>,
+        platform_version: &PlatformVersion,
     ) -> Result<Self, Error> {
         let dialect: MySqlDialect = MySqlDialect {};
         let statements: Vec<Statement> = Parser::parse_sql(&dialect, sql_string)
@@ -1110,7 +1137,8 @@ impl<'a> DriveDocumentQuery<'a> {
             )?;
         }
 
-        let internal_clauses = InternalClauses::extract_from_clauses(all_where_clauses)?;
+        let internal_clauses =
+            InternalClauses::extract_from_clauses(all_where_clauses, platform_version)?;
 
         let start_at_option = None; //todo
         let start_after_option = None; //todo
@@ -1190,6 +1218,52 @@ impl<'a> DriveDocumentQuery<'a> {
         }
     }
 
+    #[cfg(any(feature = "server", feature = "verify"))]
+    /// Versioned preflight over the non-primary-key `In` clause shape.
+    ///
+    /// Runs before any cursor storage lookup or proof processing so the
+    /// rejection precedence matches each protocol version's contract: v0
+    /// rejects more than one `In` clause with `MultipleInClauses` before a
+    /// `startAt`/`startAfter` document is ever fetched (matching the
+    /// pre-protocol-version-14 parse-time rejection), and v1 rejects the
+    /// unsupported multi-`In` + cursor combination with `Unsupported`
+    /// before spending state or proof work on the cursor. The lowering
+    /// keeps equivalent guards for callers that reach it directly.
+    pub fn validate_in_clause_shape(
+        &self,
+        platform_version: &PlatformVersion,
+    ) -> Result<(), Error> {
+        match platform_version
+            .drive
+            .methods
+            .document
+            .query
+            .non_primary_key_path_query
+        {
+            0 => {
+                if self.internal_clauses.in_clauses.len() > 1 {
+                    return Err(Error::Query(QuerySyntaxError::MultipleInClauses(
+                        "There should only be one in clause",
+                    )));
+                }
+                Ok(())
+            }
+            1 => {
+                if self.internal_clauses.in_clauses.len() > 1 && self.start_at.is_some() {
+                    return Err(Error::Query(QuerySyntaxError::Unsupported(
+                        "startAt/startAfter is not supported with multiple in clauses".to_string(),
+                    )));
+                }
+                Ok(())
+            }
+            version => Err(Error::Drive(DriveError::UnknownVersionMismatch {
+                method: "DriveDocumentQuery::validate_in_clause_shape".to_string(),
+                known_versions: vec![0, 1],
+                received: version,
+            })),
+        }
+    }
+
     #[cfg(feature = "server")]
     /// Operations to construct a path query.
     pub fn construct_path_query_operations(
@@ -1200,6 +1274,7 @@ impl<'a> DriveDocumentQuery<'a> {
         drive_operations: &mut Vec<LowLevelDriveOperation>,
         platform_version: &PlatformVersion,
     ) -> Result<PathQuery, Error> {
+        self.validate_in_clause_shape(platform_version)?;
         let drive_version = &platform_version.drive;
         // First we should get the overall document_type_path
         let document_type_path = self
@@ -1304,6 +1379,7 @@ impl<'a> DriveDocumentQuery<'a> {
         starts_at_document: Option<Document>,
         platform_version: &PlatformVersion,
     ) -> Result<PathQuery, Error> {
+        self.validate_in_clause_shape(platform_version)?;
         // First we should get the overall document_type_path
         let document_type_path = self
             .contract
@@ -1567,131 +1643,6 @@ impl<'a> DriveDocumentQuery<'a> {
     }
 
     #[cfg(any(feature = "server", feature = "verify"))]
-    /// Finds the best index for a query with more than one `In` clause, and
-    /// returns it together with the query's `In` clauses ordered by their
-    /// position in that index and the length of the equality prefix.
-    ///
-    /// A candidate index conforms when its properties decompose, left to
-    /// right, into: the equality-clause fields (exactly covering positions
-    /// `0..E`), then every `In` field on consecutive positions `E..E+K`,
-    /// then — when a range clause is present — the range field. The usual
-    /// tail rule from [`Index::matches`] still applies with the deepest `In`
-    /// field playing the role of the in field, as do the order-by continuity
-    /// rules and [`defaults::MAX_INDEX_DIFFERENCE`].
-    fn find_best_index_for_multiple_in_clauses(
-        &self,
-    ) -> Result<(&Index, Vec<&WhereClause>, usize), Error> {
-        let equal_clauses = &self.internal_clauses.equal_clauses;
-        let in_clauses = &self.internal_clauses.in_clauses;
-        let range_field = self
-            .internal_clauses
-            .range_clause
-            .as_ref()
-            .map(|range_clause| range_clause.field.as_str());
-
-        let mut fields = equal_clauses
-            .keys()
-            .map(|s| s.as_str())
-            .collect::<Vec<&str>>();
-        if let Some(range_field) = range_field {
-            fields.push(range_field);
-        }
-        fields.extend(in_clauses.iter().map(|in_clause| in_clause.field.as_str()));
-
-        let order_by_keys: Vec<&str> = self
-            .order_by
-            .keys()
-            .map(|key: &String| {
-                let str = key.as_str();
-                if !fields.contains(&str) {
-                    fields.push(str);
-                }
-                str
-            })
-            .collect();
-
-        let equality_len = equal_clauses.len();
-        let mut best: Option<(&Index, Vec<&WhereClause>, u16)> = None;
-        for index in self.document_type.indexes().values() {
-            let mut positioned: Vec<(usize, &WhereClause)> = Vec::with_capacity(in_clauses.len());
-            for in_clause in in_clauses {
-                match index
-                    .properties
-                    .iter()
-                    .position(|property| property.name == in_clause.field)
-                {
-                    Some(position) => positioned.push((position, in_clause)),
-                    None => break,
-                }
-            }
-            if positioned.len() != in_clauses.len() {
-                continue;
-            }
-            positioned.sort_by_key(|(position, _)| *position);
-
-            // The equality clauses must exactly cover the index prefix
-            if index.properties.len() < equality_len
-                || !index.properties[..equality_len]
-                    .iter()
-                    .all(|property| equal_clauses.contains_key(property.name.as_str()))
-            {
-                continue;
-            }
-            // The in clauses must sit on consecutive properties right after it
-            let consecutive_after_prefix = positioned
-                .iter()
-                .enumerate()
-                .all(|(i, (position, _))| *position == equality_len + i);
-            if !consecutive_after_prefix {
-                continue;
-            }
-            // A range clause must sit immediately after the in block
-            if let Some(range_field) = range_field {
-                match index.properties.get(equality_len + positioned.len()) {
-                    Some(property) if property.name == range_field => {}
-                    _ => continue,
-                }
-            }
-
-            let deepest_in_field = positioned
-                .last()
-                .expect("more than one in clause")
-                .1
-                .field
-                .as_str();
-            let Some(difference) = index.matches(&fields, Some(deepest_in_field), &order_by_keys)
-            else {
-                continue;
-            };
-            let ordered = positioned
-                .into_iter()
-                .map(|(_, in_clause)| in_clause)
-                .collect::<Vec<&WhereClause>>();
-            if difference == 0 {
-                return Ok((index, ordered, equality_len));
-            }
-            match &best {
-                Some((_, _, best_difference)) if *best_difference <= difference => {}
-                _ => best = Some((index, ordered, difference)),
-            }
-        }
-
-        let (index, ordered, difference) = best.ok_or(Error::Query(
-            QuerySyntaxError::WhereClauseOnNonIndexedProperty(format!(
-                "query with multiple in clauses must be for valid indexes with the in clauses on \
-                 consecutive index properties after the equality clauses, valid indexes are: {:?}",
-                self.document_type.indexes()
-            )),
-        ))?;
-        if difference > defaults::MAX_INDEX_DIFFERENCE {
-            return Err(Error::Query(QuerySyntaxError::QueryTooFarFromIndex(
-                "query must better match an existing index",
-            )));
-        }
-        Ok((index, ordered, equality_len))
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
     /// Returns a `QueryItem` given a start key and query direction.
     pub fn query_item_for_starts_at_key(starts_at_key: Vec<u8>, left_to_right: bool) -> QueryItem {
         if left_to_right {
@@ -1699,515 +1650,6 @@ impl<'a> DriveDocumentQuery<'a> {
         } else {
             QueryItem::RangeTo(..starts_at_key)
         }
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// Returns a `Query` that either starts at or after the given document ID if given.
-    fn inner_query_from_starts_at_for_id(
-        starts_at_document: Option<&StartAtDocument>,
-        left_to_right: bool,
-    ) -> Query {
-        // We only need items after the start at document
-        let mut inner_query = Query::new_with_direction(left_to_right);
-
-        if let Some(StartAtDocument {
-            document, included, ..
-        }) = starts_at_document
-        {
-            let start_at_key = document.id().to_vec();
-            if *included {
-                inner_query.insert_range_from(start_at_key..)
-            } else {
-                inner_query.insert_range_after(start_at_key..)
-            }
-        } else {
-            // No starts at document, take all NULL items
-            inner_query.insert_all();
-        }
-        inner_query
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// Returns a `Query` that either starts at or after the given key.
-    fn inner_query_starts_from_key(
-        start_at_key: Option<Vec<u8>>,
-        left_to_right: bool,
-        included: bool,
-    ) -> Query {
-        // We only need items after the start at document
-        let mut inner_query = Query::new_with_direction(left_to_right);
-
-        if left_to_right {
-            if let Some(start_at_key) = start_at_key {
-                if included {
-                    inner_query.insert_range_from(start_at_key..);
-                } else {
-                    inner_query.insert_range_after(start_at_key..);
-                }
-            } else {
-                inner_query.insert_all();
-            }
-        } else if included {
-            if let Some(start_at_key) = start_at_key {
-                inner_query.insert_range_to_inclusive(..=start_at_key);
-            } else {
-                inner_query.insert_key(vec![]);
-            }
-        } else if let Some(start_at_key) = start_at_key {
-            inner_query.insert_range_to(..start_at_key);
-        } else {
-            //todo: really not sure if this is correct
-            // Should investigate more
-            inner_query.insert_key(vec![]);
-        }
-
-        inner_query
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// Returns a `Query` that either starts at or after the given document if given.
-    fn inner_query_from_starts_at(
-        starts_at_document: Option<&StartAtDocument>,
-        indexed_property: &IndexProperty,
-        left_to_right: bool,
-        platform_version: &PlatformVersion,
-    ) -> Result<Query, Error> {
-        let mut inner_query = Query::new_with_direction(left_to_right);
-        if let Some(StartAtDocument {
-            document,
-            document_type,
-            included,
-        }) = starts_at_document
-        {
-            // We only need items after the start at document
-            let start_at_key = document.get_raw_for_document_type(
-                indexed_property.name.as_str(),
-                *document_type,
-                None,
-                platform_version,
-            )?;
-            // We want to get items starting at the start key
-            if let Some(start_at_key) = start_at_key {
-                if left_to_right {
-                    if *included {
-                        inner_query.insert_range_from(start_at_key..)
-                    } else {
-                        inner_query.insert_range_after(start_at_key..)
-                    }
-                } else if *included {
-                    inner_query.insert_range_to_inclusive(..=start_at_key)
-                } else {
-                    inner_query.insert_range_to(..start_at_key)
-                }
-            } else if left_to_right {
-                inner_query.insert_all();
-            } else {
-                inner_query.insert_key(vec![]);
-            }
-        } else {
-            // No starts at document, take all NULL items
-            inner_query.insert_all();
-        }
-        Ok(inner_query)
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    fn recursive_create_query(
-        left_over_index_properties: &[&IndexProperty],
-        unique: bool,
-        starts_at_document: Option<&StartAtDocument>, //for key level, included
-        indexed_property: &IndexProperty,
-        order_by: Option<&IndexMap<String, OrderClause>>,
-        platform_version: &PlatformVersion,
-    ) -> Result<Option<Query>, Error> {
-        match left_over_index_properties.split_first() {
-            None => Ok(None),
-            Some((first, left_over)) => {
-                let left_to_right = if let Some(order_by) = order_by {
-                    order_by
-                        .get(first.name.as_str())
-                        .map(|order_clause| order_clause.ascending)
-                        .unwrap_or(first.ascending)
-                } else {
-                    first.ascending
-                };
-
-                let mut inner_query = Self::inner_query_from_starts_at(
-                    starts_at_document,
-                    indexed_property,
-                    left_to_right,
-                    platform_version,
-                )?;
-                DriveDocumentQuery::recursive_insert_on_query(
-                    &mut inner_query,
-                    left_over,
-                    unique,
-                    starts_at_document,
-                    left_to_right,
-                    order_by,
-                    platform_version,
-                )?;
-                Ok(Some(inner_query))
-            }
-        }
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// Recursively queries as long as there are leftover index properties.
-    /// The in_start_at_document_sub_path_needing_conditional is interesting.
-    /// It indicates whether the start at document should be applied as a conditional
-    /// For example if we have a tree
-    /// Root
-    /// ├── model
-    /// │   ├── sedan
-    /// │   │   ├── brand_name
-    /// │   │   │   ├── Honda
-    /// │   │   │   │   ├── car_type
-    /// │   │   │   │   │   ├── Accord
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   ├── a47d2...
-    /// │   │   │   │   │   │   │   ├── e19c8...
-    /// │   │   │   │   │   │   │   └── f1a7b...
-    /// │   │   │   │   │   └── Civic
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── b65a7...
-    /// │   │   │   │   │       │   └── c43de...
-    /// │   │   │   ├── Toyota
-    /// │   │   │   │   ├── car_type
-    /// │   │   │   │   │   ├── Camry
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   └── 1a9d2...
-    /// │   │   │   │   │   └── Corolla
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── 3f7b4...
-    /// │   │   │   │   │       │   ├── 4e8fa...
-    /// │   │   │   │   │       │   └── 9b1c6...
-    /// │   ├── suv
-    /// │   │   ├── brand_name
-    /// │   │   │   ├── Ford*
-    /// │   │   │   │   ├── car_type*
-    /// │   │   │   │   │   ├── Escape*
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   ├── 102bc...
-    /// │   │   │   │   │   │   │   ├── 29f8e... <- Set After this document
-    /// │   │   │   │   │   │   │   └── 6b1a3...
-    /// │   │   │   │   │   └── Explorer
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── b2a9d...
-    /// │   │   │   │   │       │   └── f4d5c...
-    /// │   │   │   ├── Nissan
-    /// │   │   │   │   ├── car_type
-    /// │   │   │   │   │   ├── Rogue
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   ├── 5a9c3...
-    /// │   │   │   │   │   │   │   └── 7e4b9...
-    /// │   │   │   │   │   └── Murano
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── 8f6a2...
-    /// │   │   │   │   │       │   └── 9c7d4...
-    /// │   ├── truck
-    /// │   │   ├── brand_name
-    /// │   │   │   ├── Ford
-    /// │   │   │   │   ├── car_type
-    /// │   │   │   │   │   ├── F-150
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   ├── 72a3b...
-    /// │   │   │   │   │   │   │   └── 94c8e...
-    /// │   │   │   │   │   └── Ranger
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── 3f4b1...
-    /// │   │   │   │   │       │   ├── 6e7d2...
-    /// │   │   │   │   │       │   └── 8a1f5...
-    /// │   │   │   ├── Toyota
-    /// │   │   │   │   ├── car_type
-    /// │   │   │   │   │   ├── Tundra
-    /// │   │   │   │   │   │   ├── 0
-    /// │   │   │   │   │   │   │   ├── 7c9a4...
-    /// │   │   │   │   │   │   │   └── a5d1e...
-    /// │   │   │   │   │   └── Tacoma
-    /// │   │   │   │   │       ├── 0
-    /// │   │   │   │   │       │   ├── 1e7f4...
-    /// │   │   │   │   │       │   └── 6b9d3...
-    ///
-    /// let's say we are asking for suv's after 29f8e
-    /// here the * denotes the area needing a conditional
-    /// We need a conditional subquery on Ford to say only things after Ford (with Ford included)
-    /// We need a conditional subquery on Escape to say only things after Escape (with Escape included)
-    fn recursive_insert_on_query(
-        query: &mut Query,
-        left_over_index_properties: &[&IndexProperty],
-        unique: bool,
-        starts_at_document: Option<&StartAtDocument>, //for key level, included
-        default_left_to_right: bool,
-        order_by: Option<&IndexMap<String, OrderClause>>,
-        platform_version: &PlatformVersion,
-    ) -> Result<Option<Query>, Error> {
-        match left_over_index_properties.split_first() {
-            None => {
-                match unique {
-                    true => {
-                        query.set_subquery_key(vec![0]);
-
-                        // In the case things are NULL we allow to have multiple values
-                        let inner_query = Self::inner_query_from_starts_at_for_id(
-                            starts_at_document,
-                            true, //for ids we always go left to right
-                        );
-                        query.add_conditional_subquery(
-                            QueryItem::Key(b"".to_vec()),
-                            Some(vec![vec![0]]),
-                            Some(inner_query),
-                        );
-                    }
-                    false => {
-                        query.set_subquery_key(vec![0]);
-                        // we just get all by document id order ascending
-                        let full_query =
-                            Self::inner_query_from_starts_at_for_id(None, default_left_to_right);
-                        query.set_subquery(full_query);
-
-                        let inner_query = Self::inner_query_from_starts_at_for_id(
-                            starts_at_document,
-                            default_left_to_right,
-                        );
-
-                        query.add_conditional_subquery(
-                            QueryItem::Key(b"".to_vec()),
-                            Some(vec![vec![0]]),
-                            Some(inner_query),
-                        );
-                    }
-                }
-                Ok(None)
-            }
-            Some((first, left_over)) => {
-                let left_to_right = if let Some(order_by) = order_by {
-                    order_by
-                        .get(first.name.as_str())
-                        .map(|order_clause| order_clause.ascending)
-                        .unwrap_or(first.ascending)
-                } else {
-                    first.ascending
-                };
-
-                if let Some(start_at_document_inner) = starts_at_document {
-                    let StartAtDocument {
-                        document,
-                        document_type,
-                        included,
-                    } = start_at_document_inner;
-                    let start_at_key = document
-                        .get_raw_for_document_type(
-                            first.name.as_str(),
-                            *document_type,
-                            None,
-                            platform_version,
-                        )
-                        .ok()
-                        .flatten();
-
-                    // We should always include if we have left_over
-                    let non_conditional_included =
-                        !left_over.is_empty() || *included || start_at_key.is_none();
-
-                    let mut non_conditional_query = Self::inner_query_starts_from_key(
-                        start_at_key.clone(),
-                        left_to_right,
-                        non_conditional_included,
-                    );
-
-                    // We place None here on purpose, this has been well-thought-out
-                    // and should not change. The reason is that the path of the start
-                    // at document is used only on the conditional subquery and not on the
-                    // main query
-                    // for example in the following
-                    // Our query will be with $ownerId == a3f9b81c4d7e6a9f5b1c3e8a2d9c4f7b
-                    // With start after 8f2d5
-                    // We want to get from 2024-11-17T12:45:00Z
-                    // withdrawal
-                    // ├── $ownerId
-                    // │   ├── a3f9b81c4d7e6a9f5b1c3e8a2d9c4f7b
-                    // │   │   ├── $updatedAt
-                    // │   │   │   ├── 2024-11-17T12:45:00Z <- conditional subquery here
-                    // │   │   │   │   ├── status
-                    // │   │   │   │   │   ├── 0
-                    // │   │   │   │   │   │   ├── 7a9f1...
-                    // │   │   │   │   │   │   └── 4b8c3...
-                    // │   │   │   │   │   ├── 1
-                    // │   │   │   │   │   │   ├── 8f2d5... <- start after
-                    // │   │   │   │   │   │   └── 5c1e4...
-                    // │   │   │   │   │   ├── 2
-                    // │   │   │   │   │   │   ├── 2e7a9...
-                    // │   │   │   │   │   │   └── 1c8b3...
-                    // │   │   │   ├── 2024-11-18T11:25:00Z <- we want all statuses here, so normal subquery, with None as start at document
-                    // │   │   │   │   ├── status
-                    // │   │   │   │   │   ├── 0
-                    // │   │   │   │   │   │   └── 1a4f2...
-                    // │   │   │   │   │   ├── 2
-                    // │   │   │   │   │   │   ├── 3e7a9...
-                    // │   │   │   │   │   │   └── 198b4...
-                    // │   ├── b6d7e9c4a5f2b3d8e1a7c9f4b1e8a3f
-                    // │   │   ├── $updatedAt
-                    // │   │   │   ├── 2024-11-17T13:30:00Z
-                    // │   │   │   │   ├── status
-                    // │   │   │   │   │   ├── 0
-                    // │   │   │   │   │   │   ├── 6d7e2...
-                    // │   │   │   │   │   │   └── 9c7f5...
-                    // │   │   │   │   │   ├── 3
-                    // │   │   │   │   │   │   ├── 3a9b7...
-                    // │   │   │   │   │   │   └── 8e5c4...
-                    // │   │   │   │   │   ├── 4
-                    // │   │   │   │   │   │   ├── 1f7a8...
-                    // │   │   │   │   │   │   └── 2c9b3...
-                    // println!("going to call recursive_insert_on_query on non_conditional_query {} with left_over {:?}", non_conditional_query, left_over);
-                    DriveDocumentQuery::recursive_insert_on_query(
-                        &mut non_conditional_query,
-                        left_over,
-                        unique,
-                        None,
-                        left_to_right,
-                        order_by,
-                        platform_version,
-                    )?;
-
-                    DriveDocumentQuery::recursive_conditional_insert_on_query(
-                        &mut non_conditional_query,
-                        start_at_key,
-                        left_over,
-                        unique,
-                        start_at_document_inner,
-                        left_to_right,
-                        order_by,
-                        platform_version,
-                    )?;
-
-                    query.set_subquery(non_conditional_query);
-                } else {
-                    let mut inner_query = Query::new_with_direction(first.ascending);
-                    inner_query.insert_all();
-                    DriveDocumentQuery::recursive_insert_on_query(
-                        &mut inner_query,
-                        left_over,
-                        unique,
-                        starts_at_document,
-                        left_to_right,
-                        order_by,
-                        platform_version,
-                    )?;
-                    query.set_subquery(inner_query);
-                }
-                query.set_subquery_key(first.name.as_bytes().to_vec());
-                Ok(None)
-            }
-        }
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    #[allow(clippy::too_many_arguments)]
-    fn recursive_conditional_insert_on_query(
-        query: &mut Query,
-        conditional_value: Option<Vec<u8>>,
-        left_over_index_properties: &[&IndexProperty],
-        unique: bool,
-        starts_at_document: &StartAtDocument,
-        default_left_to_right: bool,
-        order_by: Option<&IndexMap<String, OrderClause>>,
-        platform_version: &PlatformVersion,
-    ) -> Result<(), Error> {
-        match left_over_index_properties.split_first() {
-            None => {
-                match unique {
-                    true => {
-                        // In the case things are NULL we allow to have multiple values
-                        let inner_query = Self::inner_query_from_starts_at_for_id(
-                            Some(starts_at_document),
-                            true, //for ids we always go left to right
-                        );
-                        query.add_conditional_subquery(
-                            QueryItem::Key(b"".to_vec()),
-                            Some(vec![vec![0]]),
-                            Some(inner_query),
-                        );
-                    }
-                    false => {
-                        let inner_query = Self::inner_query_from_starts_at_for_id(
-                            Some(starts_at_document),
-                            default_left_to_right,
-                        );
-
-                        query.add_conditional_subquery(
-                            QueryItem::Key(conditional_value.unwrap_or_default()),
-                            Some(vec![vec![0]]),
-                            Some(inner_query),
-                        );
-                    }
-                }
-            }
-            Some((first, left_over)) => {
-                let left_to_right = if let Some(order_by) = order_by {
-                    order_by
-                        .get(first.name.as_str())
-                        .map(|order_clause| order_clause.ascending)
-                        .unwrap_or(first.ascending)
-                } else {
-                    first.ascending
-                };
-
-                let StartAtDocument {
-                    document,
-                    document_type,
-                    ..
-                } = starts_at_document;
-
-                let lower_start_at_key = document
-                    .get_raw_for_document_type(
-                        first.name.as_str(),
-                        *document_type,
-                        None,
-                        platform_version,
-                    )
-                    .ok()
-                    .flatten();
-
-                // We include it if we are not unique,
-                // or if we are unique but the value is empty
-                let non_conditional_included = !unique || lower_start_at_key.is_none();
-
-                let mut non_conditional_query = Self::inner_query_starts_from_key(
-                    lower_start_at_key.clone(),
-                    left_to_right,
-                    non_conditional_included,
-                );
-
-                DriveDocumentQuery::recursive_insert_on_query(
-                    &mut non_conditional_query,
-                    left_over,
-                    unique,
-                    None,
-                    left_to_right,
-                    order_by,
-                    platform_version,
-                )?;
-
-                DriveDocumentQuery::recursive_conditional_insert_on_query(
-                    &mut non_conditional_query,
-                    lower_start_at_key,
-                    left_over,
-                    unique,
-                    starts_at_document,
-                    left_to_right,
-                    order_by,
-                    platform_version,
-                )?;
-
-                query.add_conditional_subquery(
-                    QueryItem::Key(conditional_value.unwrap_or_default()),
-                    Some(vec![first.name.as_bytes().to_vec()]),
-                    Some(non_conditional_query),
-                );
-            }
-        }
-        Ok(())
     }
 
     #[cfg(any(feature = "server", feature = "verify"))]
@@ -2246,425 +1688,6 @@ impl<'a> DriveDocumentQuery<'a> {
                 received: version,
             })),
         }
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// v1 of the non-primary-key path query lowering (protocol version 14):
-    /// accepts multiple `In` clauses. Single-`In` shapes lower exactly as v0.
-    fn get_non_primary_key_path_query_v1(
-        &self,
-        document_type_path: Vec<Vec<u8>>,
-        starts_at_document: Option<(Document, bool)>,
-        platform_version: &PlatformVersion,
-    ) -> Result<PathQuery, Error> {
-        if self.internal_clauses.in_clauses.len() > 1 {
-            self.get_non_primary_key_multiple_in_path_query(
-                document_type_path,
-                starts_at_document,
-                platform_version,
-            )
-        } else {
-            self.get_non_primary_key_path_query_v0(
-                document_type_path,
-                starts_at_document,
-                platform_version,
-            )
-        }
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// Lowers a query with multiple `In` clauses into a path query whose
-    /// levels carry one key set per `In` clause, in index property order,
-    /// followed by an optional range level and the usual left-over /
-    /// terminal levels. Only reachable through the v1 lowering.
-    fn get_non_primary_key_multiple_in_path_query(
-        &self,
-        document_type_path: Vec<Vec<u8>>,
-        starts_at_document: Option<(Document, bool)>,
-        platform_version: &PlatformVersion,
-    ) -> Result<PathQuery, Error> {
-        // Conservative v1: the cross-branch cursor machinery is not wired
-        // for key-set branching at more than one level, so reject cursors
-        // instead of shipping silently wrong pagination.
-        if starts_at_document.is_some() || self.start_at.is_some() {
-            return Err(Error::Query(QuerySyntaxError::Unsupported(
-                "startAt/startAfter is not supported with multiple in clauses".to_string(),
-            )));
-        }
-
-        let (index, ordered_in_clauses, equality_len) =
-            self.find_best_index_for_multiple_in_clauses()?;
-
-        // Bound the branch enumeration: the product of the in list sizes is
-        // the number of index subtrees the query opens.
-        let mut cross_product: usize = 1;
-        for in_clause in &ordered_in_clauses {
-            let in_values = in_clause.in_values().into_data_with_error()??;
-            cross_product = cross_product.saturating_mul(in_values.len());
-        }
-        if cross_product > defaults::MAX_IN_CROSS_PRODUCT_SIZE {
-            return Err(Error::Query(QuerySyntaxError::InvalidInClause(format!(
-                "the product of in clause list sizes must be at most {}, got {}",
-                defaults::MAX_IN_CROSS_PRODUCT_SIZE,
-                cross_product
-            ))));
-        }
-
-        let left_over_index_properties = index
-            .properties
-            .iter()
-            .filter(|field| {
-                !(self
-                    .internal_clauses
-                    .equal_clauses
-                    .contains_key(field.name.as_str())
-                    || ordered_in_clauses
-                        .iter()
-                        .any(|in_clause| in_clause.field == field.name)
-                    || self
-                        .internal_clauses
-                        .range_clause
-                        .as_ref()
-                        .is_some_and(|range_clause| range_clause.field == field.name))
-            })
-            .collect::<Vec<&IndexProperty>>();
-
-        // Every level that fans out (each in clause, and the range clause)
-        // needs an explicit ordering, like any range-class clause.
-        let direction_for = |field: &str| -> Result<bool, Error> {
-            let order_clause: &OrderClause = self.order_by.get(field).ok_or(Error::Query(
-                QuerySyntaxError::MissingOrderByForRange(
-                    "query must have an orderBy field for each range element",
-                ),
-            ))?;
-            Ok(order_clause.ascending)
-        };
-
-        // Build the query bottom-up. The deepest clause level is the range
-        // clause when present, otherwise the last in clause; the left-over
-        // index properties and the terminal document level hang under it.
-        let (mut child_field, mut child_query, deepest_left_to_right) =
-            match &self.internal_clauses.range_clause {
-                Some(range_clause) => {
-                    let left_to_right = direction_for(range_clause.field.as_str())?;
-                    let query = range_clause.to_path_query(
-                        self.document_type,
-                        &None,
-                        left_to_right,
-                        platform_version,
-                    )?;
-                    (range_clause.field.clone(), query, left_to_right)
-                }
-                None => {
-                    let deepest_in_clause =
-                        *ordered_in_clauses.last().expect("more than one in clause");
-                    let left_to_right = direction_for(deepest_in_clause.field.as_str())?;
-                    let query = deepest_in_clause.to_path_query(
-                        self.document_type,
-                        &None,
-                        left_to_right,
-                        platform_version,
-                    )?;
-                    (deepest_in_clause.field.clone(), query, left_to_right)
-                }
-            };
-        Self::recursive_insert_on_query(
-            &mut child_query,
-            left_over_index_properties.as_slice(),
-            index.unique,
-            None,
-            deepest_left_to_right,
-            Some(&self.order_by),
-            platform_version,
-        )?;
-
-        // Wrap the remaining in levels around it, deepest first. When a
-        // range clause is present every in clause wraps; otherwise the
-        // deepest in clause is already the leaf built above.
-        let wrapped_in_clauses = if self.internal_clauses.range_clause.is_some() {
-            ordered_in_clauses.as_slice()
-        } else {
-            &ordered_in_clauses[..ordered_in_clauses.len() - 1]
-        };
-        for in_clause in wrapped_in_clauses.iter().rev() {
-            let left_to_right = direction_for(in_clause.field.as_str())?;
-            let mut query = in_clause.to_path_query(
-                self.document_type,
-                &None,
-                left_to_right,
-                platform_version,
-            )?;
-            query.set_subquery_key(child_field.as_bytes().to_vec());
-            query.set_subquery(child_query);
-            child_field = in_clause.field.clone();
-            child_query = query;
-        }
-
-        let intermediate_values = index.properties[..equality_len]
-            .iter()
-            .map(|field| {
-                let where_clause = self
-                    .internal_clauses
-                    .equal_clauses
-                    .get(field.name.as_str())
-                    .expect("equality prefix was validated during index selection");
-                self.document_type.serialize_value_for_key(
-                    field.name.as_str(),
-                    &where_clause.value,
-                    platform_version,
-                )
-            })
-            .collect::<Result<Vec<Vec<u8>>, ProtocolError>>()
-            .map_err(Error::from)?;
-
-        let mut path = document_type_path;
-        for (intermediate_index, intermediate_value) in index.properties[..equality_len]
-            .iter()
-            .zip(intermediate_values.iter())
-        {
-            path.push(intermediate_index.name.as_bytes().to_vec());
-            path.push(intermediate_value.as_slice().to_vec());
-        }
-        path.push(child_field.as_bytes().to_vec());
-
-        Ok(PathQuery::new(
-            path,
-            SizedQuery::new(child_query, self.limit, self.offset),
-        ))
-    }
-
-    #[cfg(any(feature = "server", feature = "verify"))]
-    /// v0 of the non-primary-key path query lowering: at most one `In`
-    /// clause per query, the behavior every protocol version up to 13
-    /// committed to.
-    fn get_non_primary_key_path_query_v0(
-        &self,
-        document_type_path: Vec<Vec<u8>>,
-        starts_at_document: Option<(Document, bool)>,
-        platform_version: &PlatformVersion,
-    ) -> Result<PathQuery, Error> {
-        if self.internal_clauses.in_clauses.len() > 1 {
-            return Err(Error::Query(QuerySyntaxError::MultipleInClauses(
-                "There should only be one in clause",
-            )));
-        }
-        let index = self.find_best_index(platform_version)?;
-        let ordered_clauses: Vec<&WhereClause> = index
-            .properties
-            .iter()
-            .filter_map(|field| self.internal_clauses.equal_clauses.get(field.name.as_str()))
-            .collect();
-        let (last_clause, last_clause_is_range, subquery_clause) =
-            match self.internal_clauses.in_clauses.first() {
-                None => match &self.internal_clauses.range_clause {
-                    None => (ordered_clauses.last().copied(), false, None),
-                    Some(where_clause) => (Some(where_clause), true, None),
-                },
-                Some(in_clause) => match &self.internal_clauses.range_clause {
-                    None => (Some(in_clause), true, None),
-                    Some(range_clause) => {
-                        // Both an `in` clause and a range clause are present.
-                        // The outer path query must operate on the field that
-                        // appears *earlier* (closer to the index root) in the
-                        // chosen index, and the other clause becomes the leaf
-                        // subquery. Without this ordering, a query like
-                        // `status > 0 AND transactionIndex in [..]` on an index
-                        // `[status, transactionIndex]` builds a path that
-                        // terminates at the `status` subtree while the primary
-                        // query iterates `transactionIndex` keys, silently
-                        // returning []. See issue #2409.
-                        let position_of = |field: &str| -> Option<usize> {
-                            index
-                                .properties
-                                .iter()
-                                .position(|p| p.name.as_str() == field)
-                        };
-                        let in_pos = position_of(in_clause.field.as_str());
-                        let range_pos = position_of(range_clause.field.as_str());
-                        match (in_pos, range_pos) {
-                            (Some(i), Some(r)) if i > r => {
-                                (Some(range_clause), true, Some(in_clause))
-                            }
-                            _ => (Some(in_clause), true, Some(range_clause)),
-                        }
-                    }
-                },
-            };
-
-        // We need to get the terminal indexes unused by clauses.
-        let left_over_index_properties = index
-            .properties
-            .iter()
-            .filter(|field| {
-                !(self
-                    .internal_clauses
-                    .equal_clauses
-                    .contains_key(field.name.as_str())
-                    || (last_clause.is_some() && last_clause.unwrap().field == field.name)
-                    || (subquery_clause.is_some() && subquery_clause.unwrap().field == field.name))
-            })
-            .collect::<Vec<&IndexProperty>>();
-
-        let intermediate_values = index
-            .properties
-            .iter()
-            .filter_map(|field| {
-                match self.internal_clauses.equal_clauses.get(field.name.as_str()) {
-                    None => None,
-                    Some(where_clause) => {
-                        if !last_clause_is_range
-                            && last_clause.is_some()
-                            && last_clause.unwrap().field == field.name
-                        {
-                            //there is no need to give an intermediate value as the last clause is an equality
-                            None
-                        } else {
-                            Some(self.document_type.serialize_value_for_key(
-                                field.name.as_str(),
-                                &where_clause.value,
-                                platform_version,
-                            ))
-                        }
-                    }
-                }
-            })
-            .collect::<Result<Vec<Vec<u8>>, ProtocolError>>()
-            .map_err(Error::from)?;
-
-        let final_query = match last_clause {
-            None => {
-                // There is no last_clause which means we are using an index most likely because of an order_by, however we have no
-                // clauses, in this case we should use the first value of the index.
-                let first_index = index.properties.first().ok_or(Error::Drive(
-                    DriveError::CorruptedContractIndexes("index must have properties".to_string()),
-                ))?; // Index must have properties
-                Self::recursive_create_query(
-                    left_over_index_properties.as_slice(),
-                    index.unique,
-                    starts_at_document
-                        .map(|(document, included)| StartAtDocument {
-                            document,
-                            document_type: self.document_type,
-                            included,
-                        })
-                        .as_ref(),
-                    first_index,
-                    Some(&self.order_by),
-                    platform_version,
-                )?
-                .expect("Index must have left over properties if no last clause")
-            }
-            Some(where_clause) => {
-                let left_to_right = if where_clause.operator.is_range() {
-                    let order_clause: &OrderClause = self
-                        .order_by
-                        .get(where_clause.field.as_str())
-                        .ok_or(Error::Query(QuerySyntaxError::MissingOrderByForRange(
-                            "query must have an orderBy field for each range element",
-                        )))?;
-
-                    order_clause.ascending
-                } else {
-                    true
-                };
-
-                // We should set the starts at document to be included for the query if there are
-                // left over index properties.
-
-                let query_starts_at_document = if left_over_index_properties.is_empty() {
-                    &starts_at_document
-                } else {
-                    &None
-                };
-
-                let mut query = where_clause.to_path_query(
-                    self.document_type,
-                    query_starts_at_document,
-                    left_to_right,
-                    platform_version,
-                )?;
-
-                match subquery_clause {
-                    None => {
-                        Self::recursive_insert_on_query(
-                            &mut query,
-                            left_over_index_properties.as_slice(),
-                            index.unique,
-                            starts_at_document
-                                .map(|(document, included)| StartAtDocument {
-                                    document,
-                                    document_type: self.document_type,
-                                    included,
-                                })
-                                .as_ref(),
-                            left_to_right,
-                            Some(&self.order_by),
-                            platform_version,
-                        )?;
-                    }
-                    Some(subquery_where_clause) => {
-                        let order_clause: &OrderClause = self
-                            .order_by
-                            .get(subquery_where_clause.field.as_str())
-                            .ok_or(Error::Query(QuerySyntaxError::MissingOrderByForRange(
-                                "query must have an orderBy field for each range element",
-                            )))?;
-                        let mut subquery = subquery_where_clause.to_path_query(
-                            self.document_type,
-                            &starts_at_document,
-                            order_clause.ascending,
-                            platform_version,
-                        )?;
-                        Self::recursive_insert_on_query(
-                            &mut subquery,
-                            left_over_index_properties.as_slice(),
-                            index.unique,
-                            starts_at_document
-                                .map(|(document, included)| StartAtDocument {
-                                    document,
-                                    document_type: self.document_type,
-                                    included,
-                                })
-                                .as_ref(),
-                            left_to_right,
-                            Some(&self.order_by),
-                            platform_version,
-                        )?;
-                        let subindex = subquery_where_clause.field.as_bytes().to_vec();
-                        query.set_subquery_key(subindex);
-                        query.set_subquery(subquery);
-                    }
-                };
-
-                query
-            }
-        };
-
-        let (intermediate_indexes, last_indexes) =
-            index.properties.split_at(intermediate_values.len());
-
-        // Now we should construct the path
-        let last_index = last_indexes.first().ok_or(Error::Query(
-            QuerySyntaxError::QueryOnDocumentTypeWithNoIndexes(
-                "document query has no index with fields",
-            ),
-        ))?;
-
-        let mut path = document_type_path;
-
-        for (intermediate_index, intermediate_value) in
-            intermediate_indexes.iter().zip(intermediate_values.iter())
-        {
-            path.push(intermediate_index.name.as_bytes().to_vec());
-            path.push(intermediate_value.as_slice().to_vec());
-        }
-
-        path.push(last_index.name.as_bytes().to_vec());
-
-        Ok(PathQuery::new(
-            path,
-            SizedQuery::new(final_query, self.limit, self.offset),
-        ))
     }
 
     #[cfg(feature = "server")]
@@ -3119,14 +2142,25 @@ mod tests {
 
         let where_cbor = cbor_serializer::serializable_value_to_cbor(&query_value, None)
             .expect("expected to serialize to cbor");
-        let query =
-            DriveDocumentQuery::from_cbor(where_cbor.as_slice(), &contract, document_type, &config)
-                .expect("deserialize cbor shouldn't fail");
+        let query = DriveDocumentQuery::from_cbor(
+            where_cbor.as_slice(),
+            &contract,
+            document_type,
+            &config,
+            PlatformVersion::latest(),
+        )
+        .expect("deserialize cbor shouldn't fail");
 
         let cbor = query.to_cbor().expect("should serialize cbor");
 
-        let deserialized = DriveDocumentQuery::from_cbor(&cbor, &contract, document_type, &config)
-            .expect("should deserialize cbor");
+        let deserialized = DriveDocumentQuery::from_cbor(
+            &cbor,
+            &contract,
+            document_type,
+            &config,
+            PlatformVersion::latest(),
+        )
+        .expect("should deserialize cbor");
 
         assert_eq!(query, deserialized);
 
@@ -3160,6 +2194,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("all ranges must be on same field");
     }
@@ -3189,6 +2224,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("fields of queries must of defined supported types (where, limit, orderBy...)");
     }
@@ -3219,6 +2255,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("the query should not be created");
     }
@@ -3249,6 +2286,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("the query should be created");
     }
@@ -3278,6 +2316,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("query should be fine for a 255 byte long string");
     }
@@ -3378,6 +2417,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("fields of queries length must be under 256 bytes long");
         query
@@ -3467,6 +2507,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("The query itself should be valid for a null type");
         query
@@ -3501,6 +2542,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("query should be valid for empty array");
 
@@ -3540,6 +2582,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("query is valid for too many elements");
 
@@ -3579,6 +2622,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect("the query should be created");
 
@@ -3611,6 +2655,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("starts with can not start with an empty string");
     }
@@ -3639,6 +2684,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("starts with can not start with an empty string");
     }
@@ -3667,6 +2713,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("starts with can not start with an empty string");
     }
@@ -3695,6 +2742,7 @@ mod tests {
             &contract,
             document_type,
             &DriveConfig::default(),
+            PlatformVersion::latest(),
         )
         .expect_err("starts with can not start with an empty string");
     }
@@ -3852,8 +2900,9 @@ mod tests {
             where_clauses: Vec<WhereClause>,
             order_by_fields: &[&str],
         ) -> DriveDocumentQuery<'a> {
-            let internal_clauses = InternalClauses::extract_from_clauses(where_clauses)
-                .expect("clauses should group structurally");
+            let internal_clauses =
+                InternalClauses::extract_from_clauses(where_clauses, PlatformVersion::latest())
+                    .expect("clauses should group structurally");
             DriveDocumentQuery {
                 contract,
                 document_type: contract
@@ -3916,6 +2965,72 @@ mod tests {
                 inner.default_subquery_branch.subquery_path,
                 Some(vec![vec![0]])
             );
+        }
+
+        #[test]
+        fn descending_order_by_on_left_over_property_is_honored() {
+            let contract = family_contract();
+            let platform_version = PlatformVersion::latest();
+            // [firstName, middleName, lastName]: two in levels, lastName
+            // left over with an explicit descending order
+            let mut query = person_query(
+                &contract,
+                vec![
+                    in_clause("firstName", &["Adey", "Briney"]),
+                    in_clause("middleName", &["Ivanna", "Evangeline"]),
+                ],
+                &["firstName", "middleName"],
+            );
+            query.order_by.insert(
+                "lastName".to_string(),
+                OrderClause {
+                    field: "lastName".to_string(),
+                    ascending: false,
+                },
+            );
+
+            let path_query = query
+                .construct_path_query(None, platform_version)
+                .expect("two in clauses with a left-over order should lower");
+
+            let outer = &path_query.query.query;
+            let middle = outer
+                .default_subquery_branch
+                .subquery
+                .as_deref()
+                .expect("expected a middleName subquery");
+            assert_eq!(
+                middle.default_subquery_branch.subquery_path,
+                Some(vec![b"lastName".to_vec()])
+            );
+            let left_over_level = middle
+                .default_subquery_branch
+                .subquery
+                .as_deref()
+                .expect("expected a lastName subquery");
+            assert!(
+                !left_over_level.left_to_right,
+                "left-over lastName level must honor the descending order by"
+            );
+
+            // Without an order by entry the level falls back to the index
+            // property's direction (ascending)
+            query.order_by.shift_remove("lastName");
+            let path_query = query
+                .construct_path_query(None, platform_version)
+                .expect("two in clauses should lower");
+            let left_over_level = path_query
+                .query
+                .query
+                .default_subquery_branch
+                .subquery
+                .as_deref()
+                .expect("expected a middleName subquery")
+                .default_subquery_branch
+                .subquery
+                .as_deref()
+                .expect("expected a lastName subquery");
+            assert!(left_over_level.left_to_right);
         }
 
         #[test]
