+++ packages/rs-drive/src/verify/document/verify_start_at_document_in_proof/v0/mod.rs
@@ -37,6 +37,10 @@ impl DriveDocumentQuery<'_> {
         document_id: [u8; 32],
         platform_version: &PlatformVersion,
     ) -> Result<(RootHash, Option<Document>), Error> {
+        // Reject unsupported in-clause shapes before doing any proof work,
+        // mirroring the server-side preflight that runs before the cursor
+        // document is fetched.
+        self.validate_in_clause_shape(platform_version)?;
         let (start_at_document_path, start_at_document_key) =
             self.start_at_document_path_and_key(&document_id);
         let path_query = PathQuery::new_single_key(
